+++ src/main/java/roomescape/service/ReservationService.java
@@ -1,14 +1,53 @@
 package roomescape.service;
 
 import java.util.List;
+import org.springframework.stereotype.Service;
 import roomescape.dto.ReservationRequestDto;
 import roomescape.dto.ReservationResponseDto;
+import roomescape.entity.Reservation;
+import roomescape.repository.ReservationRepository;
+import roomescape.repository.ReservationTimeRepository;
 
-public interface ReservationService {
+@Service
+public class ReservationService {
 
-    List<ReservationResponseDto> readAll();
+    private final ReservationRepository reservationRepository;
 
-    ReservationResponseDto reserve(ReservationRequestDto reservationRequestDto);
+    private final ReservationTimeRepository reservationTimeRepository;
 
-    void cancel(Long id);
+    public ReservationService(ReservationRepository reservationRepository,
+                              ReservationTimeRepository reservationTimeRepository) {
+        this.reservationRepository = reservationRepository;
+        this.reservationTimeRepository = reservationTimeRepository;
+    }
+
+    public List<ReservationResponseDto> readAll() {
+        List<Reservation> reservations = reservationRepository.findAll();
+        return reservations.stream()
+                .map(reservation -> new ReservationResponseDto(
+                        reservation.getId(),
+                        reservation.getName(),
+                        reservation.getDate(),
+                        reservation.getTime())
+                )
+                .toList();
+    }
+
+    public ReservationResponseDto reserve(ReservationRequestDto reservationRequestDto) {
+        Reservation reservation = reservationRepository.save(new Reservation(
+                reservationRequestDto.name(),
+                reservationRequestDto.date(),
+                reservationTimeRepository.findById(reservationRequestDto.timeId())
+        ));
+        return new ReservationResponseDto(
+                reservation.getId(),
+                reservation.getName(),
+                reservation.getDate(),
+                reservation.getTime()
+        );
+    }
+
+    public void cancel(Long id) {
+        reservationRepository.delete(id);
+    }
 }
